wip: reconnect on voice server update
